backend/requirements.txt
python-dotenv
langchain
langchain-groq
python-pptx